+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - Micro-event `type` values such as event/clue/background no longer split one episode into multiple release blocks.
 # 4.0.0-alpha.1
 
+## alpha.5
+
+- Przejście do innego świata jest punktem nad wydarzeniem zamiast linią wychodzącą poza mapę.
+
 ## alpha.4
 
 - Domyślna tekstura Ziemi osadzona na izometrycznej mapie.